+++ __tests__/html/fluentTheme/side-by-side.wide.dark.js
@@ -4,5 +4,6 @@ describe('Fluent theme applied', () => {
   describe('dark theme applied', () => {
     test('side by side left - transcript, right - pre chat message', () => runHTML('fluentTheme/side-by-side.wide.dark?focus=1'));
     test('side by side left - transcript, right - transcript', () => runHTML('fluentTheme/side-by-side.wide.dark?transcript=0&transcript=2&focus=0&focus=1'));
+    test('side by side left - transcript, right - pre placeholder', () => runHTML('fluentTheme/side-by-side.wide.dark?transcript=0&transcript=3'));
   });
 });